requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
dotenv==0.9.9
python-telegram-bot==22.5